Moves input parser to UeCase.Input
