+++ src/uetools/UeCase/__init__.py
@@ -1,5 +1,6 @@
 from .Case import *
 from .Save import *
+from .Input import *
 from .Parallel import ParallelCase
 from .Config import Config
